Angular Translate Localization
A small drop-down select with a blue world icon was added to the rightmost extreme of the first page (Lookup) navigation bar.

Available languages are: 
English, 
Chinese (中文), 
Hindi (हिन्दी भाषा), 
Spanish (Castellano), 
French (Français), 
Bengali (বাংলা ভাষা), 
Russian (Русский), 
Portuguese (Português), 
Indonesian (Bahasa), 
German (Deutsche), 
Japanese (日本語), 
Swahili (Kiswahili), 
Italian (Italiano) and 
Dutch (Nederlandse)

Once a language is selected, the user can close the browser and the system will "remember" the user selection. All different views of the Explorer are instantly refreshed to display the selected language. However, a drawback is that I could not figure out how to show the translations in jQuery dialog and alert boxes, so they both will still appear in English.

The translations were originally done with php gettetx and POEDIT, widely used for localization. Then a "filter" that looks like this: {{ "Text" | translate }} was placed in the corresponding fields of the original explorer.html. explorer.js, and query-retrieve.js files (as the filter tags are required by Angular Translate). Plus, two sets of folders were added to the Explorer, one is "libs/angular" and the other one is "languages", respectively containing the required scripts and the json files used for the translations. Another new file, the angular.translate.app.js, was added to libs. 

POEDIT uses Microsoft translate, and to avoid some literal translations, some words and sentences were reviewed in Google Translate. This is my first approach to web page translations and I had a very fun time doing it, but it was not a professional job, so please do not hesitate to point out any error both in grammar or design. 

The languages used are Left to Right. If someone wants to help in Right to Left languages, and knows how to dynamically switch the .css, any help will always be well received.